# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DR TODAY API Documentation",
        "description": "API documentation for DR TODAY application",
        "contact": {
            "email": "admin@dr-today.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "DR TODAY API Server"
        }
    ],
    "paths": {
...
        "/api/user_timestamps": {
            "get": {
                "tags": [
                    "User_timestamp"
                ],
                "summary": "Get all User_timestamps",
                "operationId": "d441ec69f8804f15a0f114ff4760e5db",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "User_timestamp"
                ],
                "summary": "Create a new User_timestamps",
                "operationId": "f0def7a209ae91acd808a4cf9799ac00",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/user_timestamps/{user_timestamp}": {
            "get": {
                "tags": [
                    "User_timestamp"
                ],
                "summary": "Get all {user_timestamp}",
                "operationId": "2c048cc31986f2a889603fea57f542cc",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "put": {
                "tags": [
                    "User_timestamp"
                ],
                "summary": "Update a {user_timestamp} by ID",
                "operationId": "29d609e0fe92d76be41e28530c67ddc1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}